commit_service_test
